bzr revid: stw@openerp.com-20120831140152-9birj6mslf5g0in4

+++ openerp/addons/base/res/res_users.py
@@ -189,7 +189,7 @@ class res_users(osv.osv):
         result = super(res_users, self).read(cr, uid, ids, fields, context, load)
         canwrite = self.pool.get('ir.model.access').check(cr, uid, 'res.users', 'write', False)
         if not canwrite:
-            if isinstance(ids, (int, float)):
+            if isinstance(ids, (int, long)):
                 result = override_password(result)
             else:
                 result = map(override_password, result)
